@@ -3,14 +3,18 @@
   "main": "server.js",
   "version": "1.1.1",
   "private": true,
+  "type": "module",
+  "license": "ISC",
   "scripts": {
     "start": "node server.js",
-    "server": "nodemon server"
+    "server": "nodemon --experimental-specifier-resolution=node server",
+    "lint": "npx eslint . --ext .jsx,.js,.ts"
   },
   "description": "Connector reacted by ITGalkowski",
   "author": "Pawel Galkowski",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "bluebird": "^3.7.2",
     "body-parser": "^1.19.0",
     "chokidar": "^3.5.2",
     "concurrently": "^5.2.0",
@@ -20,7 +24,6 @@
     "core-util-is": "^1.0.3",
     "debug": "^4.3.2",
     "dotenv": "^8.6.0",
-    "eslint-config-airbnb": "^18.2.1",
     "express": "^4.17.1",
     "express-fileupload": "^1.2.1",
     "express-validator": "^6.13.0",
@@ -41,28 +44,16 @@
     "yargs": "^15.4.1"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-react": "^7.26.1",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.25.4",
     "husky": "^6.0.0",
     "lint-staged": "^11.2.3",
     "nodemon": "^2.0.13",
     "prettier": "2.3.1"
   },
-  "type":"module",
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "**/*.js": [
-      "eslint --fix",
-      "git add"
-    ]
-  },
-  "license": "ISC",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rocksdevil/apka.git"
@@ -70,6 +61,5 @@
   "bugs": {
     "url": "https://github.com/rocksdevil/apka/issues"
   },
-  "homepage": "https://github.com/rocksdevil/apka#readme",
-  "optionalDependencies": {}
+  "homepage": "https://github.com/rocksdevil/apka#readme"
 }
