@@ -1,4 +1,4 @@
-﻿export default {
+export default {
   mongoURI: 'mongodb+srv://mongouser:HEk3YPexZRP0RinZ@chat-vorap.mongodb.net/test?retryWrites=true&w=majority',
   secretOrKey: '54r5F82iFkkv8P2ps4Hda7SkpXZj1CYSAGrNheNpD0ZD',
   options: {
