@@ -1,7 +1,7 @@
 import jwt from 'jsonwebtoken';
 import config from 'config';
 
-export default function (req, res, next) {
+export default (req, res, next) => {
   const token = req.header('x-auth-token');
   if (!token) {
     return res.status(401).json({ msg: 'No token, authorization denied' });
@@ -11,8 +11,8 @@ export default function (req, res, next) {
     const decoded = jwt.verify(token, config.get('jwtSecret'));
 
     req.user = decoded.user;
-    next();
+    return next();
   } catch (err) {
-    res.status(401).json({ msg: 'Token is not valid' });
+    return res.status(401).json({ msg: 'Token is not valid' });
   }
 };
