@@ -2,41 +2,36 @@ import jwt from 'jsonwebtoken';
 import nodemailer from 'nodemailer';
 import bcrypt from 'bcryptjs';
 import dotenv from 'dotenv';
-dotenv.config()
 
-const activationMailer =  async(user, secret_key) => {
+dotenv.config();
+
+const activationMailer = async (user, secretKey) => {
   try {
-    let transporter = nodemailer.createTransport({
+    const transporter = nodemailer.createTransport({
       service: 'Gmail',
       auth: {
         user: process.env.GMAIL_USER,
         pass: process.env.GMAIL_PASS,
       },
     });
-    let aUser = user.email;
-    let userId = user.id;
+    const aUser = user.email;
+    const userId = user.id;
 
-    let msgStyle =
-      'background-color: #F0F0F0; padding: 25px 0; font-size: 15px; font-family: Georgia; line-height: 25px;';
-    let buttonStyle =
-      'padding: 10px 20px; width: 250px; height: 60px; background-color: #00BFFF; border-radius:10px; border: none; color: white; font-size: 20px; font-weight: bold; text-decoration:none;';
+    const msgStyle = 'background-color: #F0F0F0; padding: 25px 0; font-size: 15px; font-family: Georgia; line-height: 25px;';
+    const buttonStyle = 'padding: 10px 20px; width: 250px; height: 60px; background-color: #00BFFF; border-radius:10px; border: none; color: white; font-size: 20px; font-weight: bold; text-decoration:none;';
 
-    const secret_token = await bcrypt.hash(user.id, secret_key);
+    const secretToken = await bcrypt.hash(user.id, secretKey);
 
     jwt.sign(
-      {
-        user: userId,
-      },
-      secret_token,
-      {
-        expiresIn: '1d',
-      },
-      (err, secret_token) => {
-        const url = `http://localhost:3000/api/users/confirmation/${secret_token}`;
+      { user: userId },
+      secretToken,
+      { expiresIn: '1d' },
+      (err, token) => {
+        const url = `http://localhost:3000/api/users/confirmation/${token}`;
         transporter.sendMail({
           to: aUser,
           subject: 'Confirmation Email',
-          //html: `Please click this link to confirm your email: <a href="${url}">${url}</a>`
+          // html: `Please click this link to confirm your email: <a href="${url}">${url}</a>`
           html: `<div style="${msgStyle}">
           <div style="max-width: 600px; min-width: 300px: width: auto; margin: 0 auto; text-align: center;">
               <p style="text-align: left;">
@@ -56,12 +51,12 @@ const activationMailer =  async(user, secret_key) => {
           </div>
               </div>`,
         });
-      }
+      },
     );
     return true;
   } catch (err) {
     return err;
   }
-}
+};
 
 export default activationMailer;
