@@ -1,5 +1,6 @@
 import mongoose from 'mongoose';
-const Schema = mongoose.Schema;
+
+const { Schema } = mongoose;
 
 const PostSchema = new Schema({
   user: {
@@ -61,4 +62,4 @@ const PostSchema = new Schema({
 
 const Post = mongoose.model('post', PostSchema);
 
-export default Post
+export default Post;
