@@ -4,8 +4,8 @@ import jwt from 'jsonwebtoken';
 import config from 'config';
 import { check, validationResult } from 'express-validator';
 
-import auth from '../../middleware/auth.js';
-import User from '../../models/User.js';
+import auth from '../../middleware/auth';
+import User from '../../models/User';
 
 const router = express.Router();
 
@@ -14,12 +14,11 @@ router.get('/', auth, async (req, res) => {
     const user = await User.findById(req.user.id).select('-password');
     if (!user.confirmed) {
       return res.status(400).json({ errors: [{ msg: 'Verify your account first' }] });
-    } else {
-      res.json(user);
     }
+    return res.json(user);
   } catch (err) {
     console.error(err.message);
-    res.status(500).send('Server Error');
+    return res.status(500).send('Server Error');
   }
 });
 
@@ -35,7 +34,7 @@ router.post(
     const { email, password } = req.body;
 
     try {
-      let user = await User.findOne({ email });
+      const user = await User.findOne({ email });
       if (!user) {
         return res.status(400).json({ errors: [{ msg: 'Invalid Credentials' }] });
       }
@@ -65,13 +64,14 @@ router.post(
         (err, token) => {
           if (err) throw err;
           res.json({ token });
-        }
+        },
       );
+      return 'Process succes';
     } catch (err) {
       console.error(err.message);
-      res.status(500).send('Server error');
+      return res.status(500).send('Server error');
     }
-  }
+  },
 );
 
 export default router;
