@@ -1,18 +1,18 @@
-﻿import Express from 'express';
+import Express from 'express';
 import path from 'path';
 import mongoose from 'mongoose';
 import bodyParser from 'body-parser';
 import passport from 'passport';
 import fileUpload from 'express-fileupload';
-import users from './routes/api/users.js';
-import auth from './routes/api/auth.js';
-import profile from './routes/api/profile.js';
-import posts from './routes/api/posts.js';
-import forms from './routes/api/forms.js';
 import fs from 'fs';
-import db from './config/keys.js';
-import './middleware/mailer.js';
-import './config/passport.js';
+import users from './routes/api/users';
+import auth from './routes/api/auth';
+import profile from './routes/api/profile';
+import posts from './routes/api/posts';
+import forms from './routes/api/forms';
+import db from './config/keys';
+import './middleware/mailer';
+import './config/passport';
 
 const app = Express();
 
@@ -37,8 +37,8 @@ app.post('/uploads', (req, res) => {
     if (req.files === null) {
       return res.status(400).json({ msg: 'No file uploaded' });
     }
-    const file = req.files.file;
-    const user = req.body.user;
+    const { file } = req.files;
+    const { user } = req.body;
     const url = `${__dirname}/client/public/uploads/${user}/`;
 
     if (!fs.existsSync(url)) {
@@ -54,10 +54,12 @@ app.post('/uploads', (req, res) => {
         fileName: file.name,
         filePath: `/uploads/${user}/${file.name}`,
       });
+      return 'Success upload';
     });
   } catch (err) {
     res.json({ msg: 'File with that name exist' });
   }
+  return 'Process upload file finished';
 });
 
 app.post('/uploads/:company/:id', (req, res) => {
@@ -65,8 +67,8 @@ app.post('/uploads/:company/:id', (req, res) => {
     if (req.files === null) {
       return res.status(400).json({ msg: 'No file uploaded' });
     }
-    const file = req.files.file;
-    const company = req.params.company;
+    const { file } = req.files;
+    const { company } = req.params;
     const user = req.params.id;
     const url = `${__dirname}/client/public/uploads/company/${company}/${user}/`;
     if (!fs.existsSync(url)) {
@@ -82,16 +84,18 @@ app.post('/uploads/:company/:id', (req, res) => {
         fileName: file.name,
         filePath: `/uploads/company/${company}/${user}/${file.name}`,
       });
+      return 'Success upload to company';
     });
   } catch (err) {
     res.json({ msg: 'File with that name exist' });
   }
+  return 'Process upload file to company finished';
 });
 
-//Serve static assets if in production
+// Serve static assets if in production
 
 if (process.env.NODE_ENV === 'production') {
-  app.use(express.static('client/build'));
+  app.use(Express.static('client/build'));
 
   app.get('*', (req, res) => {
     res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
