@@ -1,8 +1,12 @@
-﻿const isEmpty = (value) => {
-  value === undefined ||
-    value === null ||
-    (typeof value === 'object' && Object.keys(value).length === 0) ||
-    (typeof value === 'string' && value.trim().length === 0);
+const isEmpty = (value) => {
+  if (value === undefined || value === null) {
+    return true;
+  } if (typeof value === 'object' && Object.keys(value).length === 0) {
+    return true;
+  } if (typeof value === 'string' && value.trim().length === 0) {
+    return true;
+  }
+  return false;
 };
 
-export default isEmpty
+export default isEmpty;
