@@ -1,4 +1,4 @@
-﻿import Validator from 'validator';
+import Validator from 'validator';
 import isEmpty from './is-empty';
 
 export default function validateProfileImput(data) {
@@ -58,4 +58,4 @@ export default function validateProfileImput(data) {
     errors,
     isValid: isEmpty(errors),
   };
-};
+}
